[freetype][2.12.1] corrected the resolving of dependencies
The original dependencies discovery and linking patch was incomplete,
as the logic was still dependant on *_FOUND variables, which
is redundant in case of CONFIG REQUIRED mode. Instead it should have
rely the FT_DISABLE_* variables, since those are the ones determining
whether a particular dependency/functionality should be enabled or not.