[logstalgia][1.1.4] unnecessary patch for FreeType headers
The include paths for FreeType headers should be patched,
but its CMake project sets include/freetype2/ instead of just
inclide/ for interface include directories, and so, while not being
correct, this will actually work. Ideally, it should be fixed
in the FreeType project, but no one has the time to adjust the include
paths in all of their headers.